+++ content/blog/2025-01-01-thin_delegate_aux/tests/ui/enum_dispatch/fail_panic_if_struct.rs
@@ -1,18 +1,18 @@
 #[enum_dispatch::enum_dispatch]
 trait ShapeI {
-    fn area(&self) -> usize;
+    fn area(&self) -> f64;
 }
 
 struct Rect {
-    width: usize,
-    height: usize,
+    width: f64,
+    height: f64,
 }
 
 #[enum_dispatch::enum_dispatch(ShapeI)]
 struct Shape(Rect);
 
 impl ShapeI for Rect {
-    fn area(&self) -> usize {
+    fn area(&self) -> f64 {
         self.width * self.height
     }
 }
