+++ content/blog/2025-01-01-thin_delegate_aux/tests/ui/enum_dispatch/pass_super_trait.rs
@@ -1,12 +1,12 @@
 #[enum_dispatch::enum_dispatch]
 trait ShapeI: std::fmt::Debug {
-    fn area(&self) -> usize;
+    fn area(&self) -> f64;
 }
 
 #[derive(Debug)]
 struct Rect {
-    width: usize,
-    height: usize,
+    width: f64,
+    height: f64,
 }
 
 #[derive(Debug)]
@@ -16,14 +16,14 @@ enum Shape {
 }
 
 impl ShapeI for Rect {
-    fn area(&self) -> usize {
+    fn area(&self) -> f64 {
         self.width * self.height
     }
 }
 
 fn main() {
-    let rect = Rect { width: 2, height: 3 };
-    assert_eq!(rect.area(), 6);
+    let rect = Rect { width: 2.0, height: 3.0 };
+    assert_eq!(rect.area(), 6.0);
     let shape = Shape::Rect(rect);
-    assert_eq!(shape.area(), 6);
+    assert_eq!(shape.area(), 6.0);
 }
